requirements.txt
pydantic
openai
python-dotenv
pandas
openpyxl
matplotlib 
scikit-learn
